Implemented optimistic ui / Made use of local minimongo
Here, I replaced the heavy handed Meteor.call statements with using the
client local minimongo database for faster UI.

To make this work, I had to work with user accounts, which I did by
hard-coding one. This is ok for a prototype, but will be removed, when
this becomes more mature.